requirements.txt
torch
easyocr
torchvision
supervision==0.18.0
openai==1.3.5
transformers==4.36.2
ultralytics==8.3.70
azure-identity
numpy==1.26.4
opencv-python
opencv-python-headless
gradio
dill
accelerate
timm
einops==0.8.0
paddlepaddle>=2.6.0
paddleocr==3.1.0
ruff==0.6.7
pre-commit==3.8.0
pydantic==2.10.6
pytest==8.3.3
pytest-asyncio==0.23.6
pyautogui==0.9.54
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
screeninfo
uiautomation
dashscope
groq
google-genai